backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
yfinance>=0.2.37
pandas>=2.2.1
xgboost>=2.0.3
scikit-learn>=1.4.1.post1
APScheduler>=3.10.4
requests>=2.31.0
python-dotenv>=1.0.1
pydantic>=2.6.4
